Finished make_positive but I am worried about overflows

+++ push_swap/push_swap.h
@@ -6,7 +6,7 @@
 
 typedef struct s_list
 { 
-	int			content;
+	long			content;
 	struct s_list	*next;
 }	t_list;
